+++ h2integrate/core/supported_models.py
@@ -2,6 +2,7 @@ from h2integrate.resource.river import RiverResource
 from h2integrate.core.feedstocks import FeedstockCostModel, FeedstockPerformanceModel
 from h2integrate.transporters.pipe import PipePerformanceModel
 from h2integrate.transporters.cable import CablePerformanceModel
+from h2integrate.converters.grid.grid import GridCostModel, GridPerformanceModel
 from h2integrate.finances.profast_lco import ProFastLCO
 from h2integrate.finances.profast_npv import ProFastNPV
 from h2integrate.converters.steel.steel import SteelPerformanceModel, SteelCostAndFinancialModel
@@ -21,7 +22,7 @@ from h2integrate.converters.solar.solar_pysam import PYSAMSolarPlantPerformanceM
 from h2integrate.finances.numpy_financial_npv import NumpyFinancialNPV
 from h2integrate.resource.wind.openmeteo_wind import OpenMeteoHistoricalWindResource
 from h2integrate.storage.generic_storage_cost import GenericStorageCostModel
-from h2integrate.storage.hydrogen.eco_storage import H2Storage
+from h2integrate.storage.hydrogen.mch_storage import MCHTOLStorageCostModel
 from h2integrate.converters.wind.atb_wind_cost import ATBWindPlantCostModel
 from h2integrate.storage.battery.pysam_battery import PySAMBatteryPerformanceModel
 from h2integrate.transporters.generic_combiner import GenericCombinerPerformanceModel
@@ -38,11 +39,17 @@ from h2integrate.storage.hydrogen.tank_baseclass import (
 )
 from h2integrate.converters.hydrogen.wombat_model import WOMBATElectrolyzerModel
 from h2integrate.storage.battery.atb_battery_cost import ATBBatteryCostModel
+from h2integrate.storage.hydrogen.h2_storage_cost import (
+    PipeStorageCostModel,
+    SaltCavernStorageCostModel,
+    LinedRockCavernStorageCostModel,
+)
 from h2integrate.converters.iron.humbert_ewin_perf import HumbertEwinPerformanceComponent
 from h2integrate.converters.ammonia.ammonia_synloop import (
     AmmoniaSynLoopCostModel,
     AmmoniaSynLoopPerformanceModel,
 )
+from h2integrate.storage.simple_storage_auto_sizing import StorageAutoSizingModel
 from h2integrate.converters.water.desal.desalination import (
     ReverseOsmosisCostModel,
     ReverseOsmosisPerformanceModel,
@@ -55,6 +62,8 @@ from h2integrate.converters.hydrogen.pem_electrolyzer import (
 from h2integrate.converters.solar.atb_res_com_pv_cost import ATBResComPVCostModel
 from h2integrate.converters.solar.atb_utility_pv_cost import ATBUtilityPVCostModel
 from h2integrate.resource.wind.nrel_developer_wtk_api import WTKNRELDeveloperAPIWindResource
+from h2integrate.converters.iron.martin_mine_cost_model import MartinIronMineCostComponent
+from h2integrate.converters.iron.martin_mine_perf_model import MartinIronMinePerformanceComponent
 from h2integrate.converters.methanol.smr_methanol_plant import (
     SMRMethanolPlantCostModel,
     SMRMethanolPlantFinanceModel,
@@ -79,6 +88,7 @@ from h2integrate.converters.co2.marine.direct_ocean_capture import DOCCostModel,
 from h2integrate.control.control_strategies.pyomo_controllers import (
     HeuristicLoadFollowingController,
 )
+from h2integrate.converters.hydrogen.geologic.mathur_modified import GeoH2SubsurfaceCostModel
 from h2integrate.resource.solar.nrel_developer_goes_api_models import (
     GOESTMYSolarAPI,
     GOESConusSolarAPI,
@@ -88,18 +98,17 @@ from h2integrate.resource.solar.nrel_developer_goes_api_models import (
 from h2integrate.converters.hydrogen.eco_tools_pem_electrolyzer import (
     ECOElectrolyzerPerformanceModel,
 )
-from h2integrate.control.control_strategies.openloop_controllers import (
-    DemandOpenLoopController,
-    PassThroughOpenLoopController,
-)
 from h2integrate.converters.water_power.hydro_plant_run_of_river import (
     RunOfRiverHydroCostModel,
     RunOfRiverHydroPerformanceModel,
 )
-from h2integrate.converters.hydrogen.geologic.natural_geoh2_plant import (
-    NaturalGeoH2CostModel,
-    NaturalGeoH2FinanceModel,
+from h2integrate.converters.hydrogen.geologic.simple_natural_geoh2 import (
     NaturalGeoH2PerformanceModel,
+)
+from h2integrate.resource.solar.nrel_developer_himawari_api_models import (
+    Himawari7SolarAPI,
+    Himawari8SolarAPI,
+    HimawariTMYSolarAPI,
 )
 from h2integrate.control.control_rules.converters.generic_converter import (
     PyomoDispatchGenericConverter,
@@ -112,13 +121,27 @@ from h2integrate.converters.co2.marine.ocean_alkalinity_enhancement import (
 from h2integrate.converters.hydrogen.custom_electrolyzer_cost_model import (
     CustomElectrolyzerCostModel,
 )
-from h2integrate.converters.hydrogen.geologic.stimulated_geoh2_plant import (
-    StimulatedGeoH2CostModel,
-    StimulatedGeoH2FinanceModel,
+from h2integrate.converters.hydrogen.geologic.templeton_serpentinization import (
     StimulatedGeoH2PerformanceModel,
 )
 from h2integrate.control.control_rules.storage.pyomo_storage_rule_baseclass import (
     PyomoRuleStorageBaseclass,
+)
+from h2integrate.control.control_strategies.passthrough_openloop_controller import (
+    PassThroughOpenLoopController,
+)
+from h2integrate.resource.solar.nrel_developer_meteosat_prime_meridian_models import (
+    MeteosatPrimeMeridianSolarAPI,
+    MeteosatPrimeMeridianTMYSolarAPI,
+)
+from h2integrate.control.control_strategies.storage.demand_openloop_controller import (
+    DemandOpenLoopStorageController,
+)
+from h2integrate.control.control_strategies.converters.demand_openloop_controller import (
+    DemandOpenLoopConverterController,
+)
+from h2integrate.control.control_strategies.converters.flexible_demand_openloop_controller import (
+    FlexibleDemandOpenLoopConverterController,
 )
 
 
@@ -131,6 +154,11 @@ supported_models = {
     "goes_conus_solar_v4_api": GOESConusSolarAPI,
     "goes_fulldisc_solar_v4_api": GOESFullDiscSolarAPI,
     "goes_tmy_solar_v4_api": GOESTMYSolarAPI,
+    "meteosat_solar_v4_api": MeteosatPrimeMeridianSolarAPI,
+    "meteosat_tmy_solar_v4_api": MeteosatPrimeMeridianTMYSolarAPI,
+    "himawari7_solar_v3_api": Himawari7SolarAPI,
+    "himawari8_solar_v3_api": Himawari8SolarAPI,
+    "himawari_tmy_solar_v3_api": HimawariTMYSolarAPI,
     # Converters
     "atb_wind_cost": ATBWindPlantCostModel,
     "pysam_wind_plant_performance": PYSAMWindPlantPerformanceModel,
@@ -154,6 +182,8 @@ supported_models = {
     "iron_mine_cost": IronMineCostComponent,
     "iron_plant_performance": IronPlantPerformanceComponent,
     "iron_plant_cost": IronPlantCostComponent,
+    "iron_mine_performance_martin": MartinIronMinePerformanceComponent,  # standalone model
+    "iron_mine_cost_martin": MartinIronMineCostComponent,  # standalone model
     "humbert_electrowinning_performance": HumbertEwinPerformanceComponent,
     "humbert_stinn_electrowinning_cost": HumbertStinnEwinCostComponent,
     "reverse_osmosis_desalination_performance": ReverseOsmosisPerformanceModel,
@@ -175,12 +205,9 @@ supported_models = {
     "ocean_alkalinity_enhancement_performance": OAEPerformanceModel,
     "ocean_alkalinity_enhancement_cost": OAECostModel,
     "ocean_alkalinity_enhancement_cost_financial": OAECostAndFinancialModel,
-    "natural_geoh2_performance": NaturalGeoH2PerformanceModel,
-    "natural_geoh2_cost": NaturalGeoH2CostModel,
-    "natural_geoh2": NaturalGeoH2FinanceModel,
-    "stimulated_geoh2_performance": StimulatedGeoH2PerformanceModel,
-    "stimulated_geoh2_cost": StimulatedGeoH2CostModel,
-    "stimulated_geoh2": StimulatedGeoH2FinanceModel,
+    "simple_natural_geoh2_performance": NaturalGeoH2PerformanceModel,
+    "templeton_serpentinization_geoh2_performance": StimulatedGeoH2PerformanceModel,
+    "mathur_modified_geoh2_cost": GeoH2SubsurfaceCostModel,
     "natural_gas_performance": NaturalGasPerformanceModel,
     "natural_gas_cost": NaturalGasCostModel,
     # Transport
@@ -194,27 +221,43 @@ supported_models = {
     "summer": GenericSummerPerformanceModel,
     # Storage
     "pysam_battery": PySAMBatteryPerformanceModel,
-    "h2_storage": H2Storage,
     "hydrogen_tank_performance": HydrogenTankPerformanceModel,
     "hydrogen_tank_cost": HydrogenTankCostModel,
+    "storage_auto_sizing": StorageAutoSizingModel,
+    "lined_rock_cavern_h2_storage_cost": LinedRockCavernStorageCostModel,
+    "salt_cavern_h2_storage_cost": SaltCavernStorageCostModel,
+    "mch_tol_h2_storage_cost": MCHTOLStorageCostModel,
+    "buried_pipe_h2_storage_cost": PipeStorageCostModel,
     "atb_battery_cost": ATBBatteryCostModel,
     "generic_storage_cost": GenericStorageCostModel,
     "simple_generic_storage": SimpleGenericStorage,
     # Control
     "pass_through_controller": PassThroughOpenLoopController,
-    "demand_open_loop_controller": DemandOpenLoopController,
+    "demand_open_loop_storage_controller": DemandOpenLoopStorageController,
     "heuristic_load_following_controller": HeuristicLoadFollowingController,
+    "demand_open_loop_converter_controller": DemandOpenLoopConverterController,
+    "flexible_demand_open_loop_converter_controller": FlexibleDemandOpenLoopConverterController,
     # Dispatch
     "pyomo_dispatch_generic_converter": PyomoDispatchGenericConverter,
     "pyomo_dispatch_generic_storage": PyomoRuleStorageBaseclass,
     # Feedstock
     "feedstock_performance": FeedstockPerformanceModel,
     "feedstock_cost": FeedstockCostModel,
+    # Grid
+    "grid_performance": GridPerformanceModel,
+    "grid_cost": GridCostModel,
     # Finance
     "ProFastComp": ProFastLCO,
     "ProFastNPV": ProFastNPV,
     "NumpyFinancialNPV": NumpyFinancialNPV,
 }
 
-
-electricity_producing_techs = ["wind", "solar", "pv", "river", "hopp", "natural_gas_plant"]
+electricity_producing_techs = [
+    "wind",
+    "solar",
+    "pv",
+    "river",
+    "hopp",
+    "natural_gas_plant",
+    "grid_buy",
+]